Organize helper functions into their own scope
