+++ src/commands/meta.ts
@@ -8,21 +8,9 @@
  */
 
 import * as vscode from 'vscode'
-import * as os from 'os'
-import * as fs from 'fs'
-import { execSync } from 'child_process'
 
-
-const getDistro = () => {
-    try {
-        if (process.platform === 'linux') {
-            const data = fs.readFileSync('/etc/os-release', 'utf8')
-            const m = data.match(/^PRETTY_NAME=(?:"|')?(.+?)(?:"|')?$/m)
-            if (m) return m[1]
-        }
-    } catch (e) {}
-    return ''
-}
+import { arturoVersion } from './helper/arturo'
+import { osInfo } from './helper/system'
 
 /** Opens the Arturo GitHub issues page to report a new issue.
  * 
@@ -44,13 +32,8 @@ export const reportIssue = async () => {
     const ext = vscode.extensions.getExtension('drkameleon.arturo')
     const extensionVersion = ext?.packageJSON?.version || 'unknown'
 
-    const distro = getDistro()
-    const osLine = `${os.type()} ${distro ? ' - ' + distro : ''}`
-
-    let arturoVersion = 'not found'
-    try {
-        arturoVersion = execSync('arturo --version', { encoding: 'utf8', stdio: ['ignore', 'pipe', 'ignore'] }).trim()
-    } catch (e) {
+    let arturo: string | null = arturoVersion()
+    if (arturo === null) {
         const reason = 'Arturo command line tool not found.'
         const recommendation = 'Please ensure Arturo is installed and added to your PATH.'
         vscode.window.showErrorMessage(`${reason} ${recommendation}`)
@@ -68,8 +51,8 @@ export const reportIssue = async () => {
         '...',
         '',
         '**System**',
-        `- OS: ${osLine}`,
-        `- Arturo: ${arturoVersion}`,
+        `- OS: ${osInfo()}`,
+        `- Arturo: ${arturo}`,
         `- VS Code: ${appName} ${vscodeVersion} / Extension: v${extensionVersion}`,
         ``,
         '**More Information**',
